Return 1024 from _GetMaxPacketSize
We might be able to get away with this since FTDI's buffer is 1K, but the question is whether FTDI starts operating on the data before it's received the entire USB transfer to fill the buffer. If so, it's possible for a TEST=[0,1] command series to straddle the 512 byte mark (where USB packets are split). If FTDI does start work before receiving the entire transfer, it likely operates on a USB-packet basis, and it's possible that the TEST=[0,1] command series straddles the line of the two 512-byte packets, which could trigger a 0-Volt pulse that causes SBW mode to exit (ie >7us 0V pulse)